+++ migrations/0002_eminent_thunderbolt.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "bot_responses" ADD COLUMN "parution_ids" jsonb;